requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
httpx==0.27.0
python-multipart==0.0.9
jinja2==3.1.4
python-dotenv==1.0.1
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.148.0
numpy==1.26.4
polars==1.6.0
sentence-transformers==3.0.1
